// src/platform/packages/shared/kbn-grouping/src/containers/query/types.d.ts
import type { Script, MappingRuntimeField, MappingRuntimeFields } from '@elastic/elasticsearch/lib/api/types';
import type { RuntimeFieldSpec, RuntimePrimitiveTypes } from '@kbn/data-views-plugin/common';
import type { BoolQuery } from '@kbn/es-query';
import type { estypes } from '@elastic/elasticsearch';
type RunTimeMappings = Record<string, Omit<RuntimeFieldSpec, 'type'> & {
    type: RuntimePrimitiveTypes;
}> | undefined;
export interface BoolAgg {
    bool: BoolQuery;
}
export interface RangeAgg {
    range: {
        '@timestamp': {
            gte: string;
            lte: string;
        };
    };
}
export type NamedAggregation = Record<string, estypes.AggregationsAggregationContainer>;
export type GroupingSort = Array<{
    [category: string]: {
        order: 'asc' | 'desc';
    };
}>;
export interface GroupingQueryArgs {
    additionalFilters: BoolAgg[];
    groupByField: string;
    rootAggregations?: NamedAggregation[];
    runtimeMappings?: RunTimeMappings;
    additionalAggregationsRoot?: NamedAggregation[];
    pageNumber?: number;
    uniqueValue: string;
    size?: number;
    sort?: GroupingSort;
    statsAggregations?: NamedAggregation[];
    timeRange?: {
        from: string;
        to: string;
    };
    multiValueFieldsToFlatten?: string[];
    countByKeyForMultiValueFields?: string;
    unitsCountFilter?: estypes.QueryDslQueryContainer;
}
export interface MainAggregation extends NamedAggregation {
    groupByFields: {
        aggs: NamedAggregation;
        terms: estypes.AggregationsAggregationContainer['terms'];
    };
}
export interface GroupingRuntimeField extends MappingRuntimeField {
    script: Script & {
        params: Record<string, any>;
    };
}
type GroupingMappingRuntimeFields = Record<'groupByField', GroupingRuntimeField>;
export type GroupingQuery = estypes.QueryDslQueryContainer & {
    aggs: MainAggregation;
    query: {
        bool: {
            filter: Array<BoolAgg | RangeAgg>;
        };
    };
    runtime_mappings: MappingRuntimeFields & GroupingMappingRuntimeFields;
    size: number;
    _source: boolean;
};
export {};